requirements.txt
celery==5.2.6
Django==3.2.13
django-redis==5.2.0
djangorestframework==3.13.1
pre-commit==2.19.0
